done room backend part 1

+++ admin/facilities_ajax.php
@@ -41,10 +41,17 @@
         $frm_data = filteration($_POST);
         $values = [$frm_data['rem_facilities']];
 
+        $check_q = select('SELECT * FROM `room_facilities` WHERE `facilities_id`=?',[$frm_data['rem_facilities']],'i');
 
-        $q = "DELETE FROM `features` WHERE `id`=?";
-        $res = delete($q, $values,'i');
-        echo $res;
+        if(mysqli_num_rows($check_q)==0){
+
+            $q = "DELETE FROM `features` WHERE `id`=?";
+            $res = delete($q, $values,'i');
+            echo $res;
+        }else{
+            echo 'room_added';
+        }
+
     }
 
 
@@ -100,7 +107,6 @@
         $values = [$frm_data['rem_features']];
 
      
-
         $q = "DELETE FROM `facilities` WHERE `id`=?";
         $res = delete($q, $values,'i');
         echo $res;
